% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pexm.R
\docType{package}
\name{pexm}
\alias{pexm}
\title{pexm: Loading a Just Another Gibbs Sampling Module for the Piecewise Exponential Distribution}
\description{
Load the Just Another Gibbs Sampling (JAGS) module "pexm". 
The module provides the tools to work with the Piecewise Exponential (PE) distribution in a Bayesian model with 
the corresponding Markov Chain Monte Carlo algorithm (Gibbs Sampling) implemented via JAGS.
}
\details{
The current version contains two main functions: \code{loadpexm} loads the new JAGS module. 
Details about JAGS can be found in the \href{http://sourceforge.net/projects/mcmc-jags/files/Manuals/3.x/jags_user_manual.pdf}{JAGS User Manual}.
Download instructions are presented in the home page given by this \href{http://mcmc-jags.sourceforge.net}{link}.
}
\seealso{
\code{\link{loadpexm}}
}
\author{
Vinicius D. Mayrink, Joao D. N. Duarte and Fabio N. Demarqui. 

Departamento de Estatistica, ICEx, Universidade Federal de Minas Gerais, Av. Antonio Carlos, 6627, Belo Horizonte, MG, Brazil, 31270-901.
}
